Add a homepage for finder-frontend
When visiting review apps, or spinning the app up locally, it's handy
(especially for new users) to be able to see links to some of the pages
rendered in the app.

It's quite easy to be confused by an error message when all you need to do
is magically know the URL of one of the pages accepted by the app.

This will never be seen in production because we won't publish a route
to this page on finder-frontend.