Display error message when user enters bad date
When filtering by date on a finder, if the user
enters an invalid date we don't apply that date
filter. But we also do not display an error to
inform them of their mistake. This commit adds
error message fields to the JSON that is returned
so that we can display appropriate error messaging
on the front end. That work will be added at a later
date.

Return error message status in JSON response